requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.0.1
google-generativeai
faiss-cpu
numpy